api/python/ai-server-requirements.txt
bchlib==1.0.0
Flask[async]
Flask_Cors==4.0.0
numpy==1.21.6
opencv_python==4.8.0.76
Pillow==9.5.0
tensorflow==1.13.2
protobuf==3.20.0
